Added support for building shadow model experiments from an attack config; added transforms to dataset config
Fixed "watermarks" portion of the attack schema to include "type" and "properties"

Updated the model schema to include 'tfgloro' as a framework option. Updated the gloro evaluator to send lib-gloro the correct file path type and to compile the model with the missing loss and accuracy metric. This should ideally be done through configs, but I'm not sure how to get that to work yet.

Removed shadow model evaluator: moved to project-specific repo.

Deleted tfgloro; moved lib-gloro wrappers to project-specific repo.